UI: Create overlay models and refactoring code
