+++ ui-ngx/src/app/shared/components/popover.models.ts
@@ -15,8 +15,9 @@
 ///
 
 import { animate, AnimationTriggerMetadata, style, transition, trigger } from '@angular/animations';
-import { ConnectedOverlayPositionChange, ConnectionPositionPair } from '@angular/cdk/overlay';
+import { ConnectedOverlayPositionChange } from '@angular/cdk/overlay';
 import { TbPopoverComponent } from '@shared/components/popover.component';
+import { POSITION_MAP } from '@shared/models/overlay.models';
 
 export const popoverMotion: AnimationTriggerMetadata = trigger('popoverMotion', [
   transition('void => active', [
@@ -45,34 +46,7 @@ export const PopoverPlacements = ['top', 'topLeft', 'topRight', 'right', 'rightT
 type PopoverPlacementTuple = typeof PopoverPlacements;
 export type PopoverPlacement = PopoverPlacementTuple[number];
 
-export const POSITION_MAP: { [key: string]: ConnectionPositionPair } = {
-  top: new ConnectionPositionPair({ originX: 'center', originY: 'top' }, { overlayX: 'center', overlayY: 'bottom' }),
-  topLeft: new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'start', overlayY: 'bottom' }),
-  topRight: new ConnectionPositionPair({ originX: 'end', originY: 'top' }, { overlayX: 'end', overlayY: 'bottom' }),
-  right: new ConnectionPositionPair({ originX: 'end', originY: 'center' }, { overlayX: 'start', overlayY: 'center' }),
-  rightTop: new ConnectionPositionPair({ originX: 'end', originY: 'top' }, { overlayX: 'start', overlayY: 'top' }),
-  rightBottom: new ConnectionPositionPair(
-    { originX: 'end', originY: 'bottom' },
-    { overlayX: 'start', overlayY: 'bottom' }
-  ),
-  bottom: new ConnectionPositionPair({ originX: 'center', originY: 'bottom' }, { overlayX: 'center', overlayY: 'top' }),
-  bottomLeft: new ConnectionPositionPair(
-    { originX: 'start', originY: 'bottom' },
-    { overlayX: 'start', overlayY: 'top' }
-  ),
-  bottomRight: new ConnectionPositionPair({ originX: 'end', originY: 'bottom' }, { overlayX: 'end', overlayY: 'top' }),
-  left: new ConnectionPositionPair({ originX: 'start', originY: 'center' }, { overlayX: 'end', overlayY: 'center' }),
-  leftTop: new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'end', overlayY: 'top' }),
-  leftBottom: new ConnectionPositionPair(
-    { originX: 'start', originY: 'bottom' },
-    { overlayX: 'end', overlayY: 'bottom' }
-  )
-};
-
 export const DEFAULT_POPOVER_POSITIONS = [POSITION_MAP.top, POSITION_MAP.right, POSITION_MAP.bottom, POSITION_MAP.left];
-
-export const DEFAULT_OVERLAY_POSITIONS = [POSITION_MAP.bottomLeft, POSITION_MAP.bottomRight, POSITION_MAP.topLeft,
-  POSITION_MAP.topRight, POSITION_MAP.left, POSITION_MAP.right];
 
 export function getPlacementName(position: ConnectedOverlayPositionChange): PopoverPlacement | undefined {
   for (const placement in POSITION_MAP) {
